Implement URL encoding/decoding of messages
This adds a second PublicKeyShare message to the protocol buffer
definitions as well as a 'oneof container class' that may contain any
message we're going to send to tendermint.

We also implement encoding and decoding of messages as URL compatible
strings.

Finally, there are also the first tests for the URL encoding/decoding.