Implement executor.
The executor's job is to make sure that the executor contract executes
all batches. It runs in its own goroutine and receives batch parameters
(most importantly, batch decryption keys, decrypted transactions, and
decryption signatures) via a channel. When it does not receive
parameters in a while, it attempts to skip cipher batches.

For every tx, there's a kick off block number. At the moment, this
is just the earliest number at which the contract allows sending it. In
the future, this can be easily extended to stagger execution, as an easy
way to allow keypers to distribute work among them.

It seems to work in general, but requires more testing. In particular,
the executor seems to be unable to keep up with key generation, at least
at low batch spans.