Add work-in-progress keyper2 command
This is a reimplementation of the keyper that tries to bring more
structure into our code.

We strictly separate the information we retrieve from external sources
into different structs and keep another struct for our internal keyper
state.

The keyper now only runs a loop that fetches the latest information
from shutter and geth. Then the decide.Decider class determines
actions to take, but doesn't immediately execute them, but rather
keeps a list. After the decider is done, those actions are executed in
a separate step.