Implement optimistic signature checking
Keypers are initially trusted to execute cipher batches faithfully. If
they don't, they can be challenged to provide decryption signatures from
a threshold of the keyper set justifying their action. If they don't do
that in an appropriate amount of time, they can be slashed.

To prevent griefing, only keypers can accuse other keypers. In the
future, we might want to require an additional deposit.