Allow multiple PolyEval messages from a single keyper
The new keyper may send multiple PolyEval messages when other keypers
register after the DKG process has already started, because the keyper
must wait for the encryption key.

Also, do not store the DKG messages. We do not need them in
shuttermint and the keyper can use events to query the necessary
information.