Don't wait for tx in action
Don't wait inside action.Run for Ethereum txs to be included (which can
take a long time). Instead, just send them and log when they get
included.

Since we don't wait for the tx to have an effect, we have to make sure
that we don't send the same tx again in the next iteration. This could
be done by storing the tx in the state in a list of pending txs.
However, the tx is created inside an action, but state can only be
accessed in the decider. Therefore, as an alternative, we store "what"
we wanted to do and check if this is done (either by our tx or by
someone else). This is not ideal though.