Return error if main chain disconnects
Prior, we would log an error, but keep running. Now, the run function
returns an error. The same is not done yet for the shutter chain.

Also, force a sync of both chains on start. Otherwise, one might be much
more recent than the other one. This is especially a problem on fresh
runs when one chain has not been synced at all.