Add config script
Adds a config script that can be used to schedule and query configs.
Configs are passed around as JSON files. Scheduling is broken up into
two parts: Setting the next values and only then sending the scheduling
transaction. This allows double checking that the values are correct.

Configs on the chain can be checked with `query` and `list`.

Parts of this can already be done with the prepare script. However, the
latter is intended for preparing test setups and requires keyper configs
to exist.