Refactor half step execution
Split maybeExecuteHalfStep. Handle the cases where we're not a keyper,
and only can skip the cipher execution after the timeout.

The handling of the execution delay is still fishy. We don't have a
delay for plain batch execution. That delay should start after the
cipher execution has finished, but we would need to look at the main chain
to determine the block when this happens.