Rewrite key broadcasting contract
The old version expected signatures to be collected offchain and
submitted once. In the new version, every keyper sends their own
transactions. This is acceptable, because we only broadcast eon keys
which happens very infrequently.

The contract ensures that every keyper only votes once per start batch
index (i.e., eon). It does not try to announce a winner.

It is possible for keypers to vote for start batch indices very far into
the future. If the keyper set changes in the mean time, the new keypers
would not be able to vote again. In the unlikely case that this happens,
this can be mitigated by simply deploying a new key broadcasting
contract when a the new keyper set is scheduled.