Work around block number polling issue
Polling the block number in Metamask does not work very well as they
cache it for a long time, so often results are outdated.

To work around this issue, we now query the current block time only
infrequently (25s) to ensure we always get an up to date one. To compute
the current block, we use the latest queried block number and the time
that passed since the query, assuming a fixed known block time.