Allow adding txs to next batch
Prior, you could only get your txs included in the current batch. Now,
you can do that to the current and the next one, so there are two open
batches at all times.

This also changes the way the config change heads up works: It is now
guaranteed to be at least one batch in order to assure that one can't
change the config for a batch that's already accepting txs.

Ideally, we should have another parameter "batch interval" in addition
to "batch span". But the current solution is easier for now
Effectively, the batch interval is the batch span and the batch span is
twice the batch span (and our terminology doesn't make sense anymore).